Remove deprecated session detail and chat components; refactor imports for UserInputCharacterButton. Clean up unused files and update related references.
