+++ apps/pwa/src/pages/sessions/ui/chat/index.ts
@@ -1,6 +1,5 @@
-export { default as SessionContent } from "./session-content";
 export { MessageItemInternal } from "./message-components";
-export { default as UserInputs } from "./user-inputs";
+export { UserInputCharacterButton } from "./user-input-character-button";
 export { SessionSettings } from "./session-settings";
 export { InlineChatStyles } from "./inline-chat-styles";
